% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpa.r
\name{lcpa_cplex}
\alias{lcpa_cplex}
\title{Lattice Cutting Plane Algorithm using CPLEX}
\usage{
lcpa_cplex(x, y, logfile, R_max, time_limit)
}
\arguments{
\item{x}{feature matrix}

\item{y}{response vector}

\item{logfile}{file to write log output}

\item{R_max}{max number of features (including intercept)}

\item{time_limit}{max running time in seconds}
}
\description{
Lattice Cutting Plane Algorithm using CPLEX
}
