% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskslimr.cplex-package.R
\docType{package}
\name{riskslimr.cplex-package}
\alias{riskslimr.cplex}
\alias{riskslimr.cplex-package}
\title{riskslimr.cplex: CPLEX backend for riskslimr}
\description{
GLPK backend for riskslimr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zamorarr/riskslimr.cplex}
  \item Report bugs at \url{https://github.com/zamorarr/riskslimr.cplex/issues}
}

}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
