% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskslimr.cplex-package.R
\docType{package}
\name{riskslimr.cplex-package}
\alias{riskslimr.cplex}
\alias{riskslimr.cplex-package}
\title{riskslimr.cplex: CPLEX backend for riskslimr}
\description{
GLPK backend for riskslimr
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
